+++ app/lib/ui/components/navigation/navigation_bar_customization.dart
@@ -98,7 +98,7 @@ class ItemBadgeState {
 
   List<ItemBadge> _itemBadgeList = [
     ItemBadge.none,
-    ItemBadge.standard,
+    ItemBadge.dot,
     ItemBadge.count,
   ];